Re #26957 Remove SANS Convert To Q Workflw. Alg
Removes the SANSConvertToQ Workflow algorithm as part of the SANS
history changes. It has been converted to a class
