+++ Framework/PythonInterface/plugins/algorithms/WorkflowAlgorithms/SANS/SANSBeamCentreFinderCore.py
@@ -12,6 +12,7 @@ from __future__ import (absolute_import, division, print_function)
 from mantid.api import (DataProcessorAlgorithm, MatrixWorkspaceProperty, AlgorithmFactory, PropertyMode, Progress,
                         IEventWorkspace)
 from mantid.kernel import (Direction, PropertyManagerProperty, StringListValidator)
+from sans.algorithm_detail.ConvertToQ import ConvertToQ
 from sans.common.constants import EMPTY_NAME
 from sans.common.general_functions import create_child_algorithm, append_to_sans_file_tag
 from sans.state.state_base import create_deserialized_sans_state_from_property_manager
@@ -75,8 +76,8 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         self.declareProperty("Centre1", 0.0, direction=Direction.Input)
         self.declareProperty("Centre2", 0.0, direction=Direction.Input)
 
-        self.declareProperty("RMax", 0.26, direction= Direction.Input)
-        self.declareProperty("RMin", 0.06, direction= Direction.Input)
+        self.declareProperty("RMax", 0.26, direction=Direction.Input)
+        self.declareProperty("RMin", 0.06, direction=Direction.Input)
 
         # ----------
         # OUTPUT
@@ -112,7 +113,7 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         state.mask.use_mask_phi_mirror = True
 
         # Set compatibility mode
-        #state.compatibility.use_compatibility_mode = self.getProperty('CompatibilityMode').value
+        # state.compatibility.use_compatibility_mode = self.getProperty('CompatibilityMode').value
 
         # Set test centre
         state.move.detectors[DetectorType.to_string(DetectorType.LAB)].sample_centre_pos1 = \
@@ -244,11 +245,14 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
             xml_string = quadrant_xml(centre, r_min, r_max, quadrant)
             quadrant_scatter_data = self._get_cloned_workspace(scatter_data)
             self._mask_quadrants(quadrant_scatter_data, xml_string)
-            quadrant_scatter_data, sum_of_counts, sum_of_norms = self._convert_to_q(state_serialized,
-                                                                                    quadrant_scatter_data,
-                                                                                    wavelength_adjustment_workspace,
-                                                                                    pixel_adjustment_workspace,
-                                                                                    wavelength_and_pixel_adjustment_workspace)
+
+            quadrant_scatter_data, sum_of_counts, sum_of_norms = \
+                self._convert_to_q(state=state,
+                                   workspace=quadrant_scatter_data,
+                                   wavelength_adjustment_workspace=wavelength_adjustment_workspace,
+                                   pixel_adjustment_workspace=pixel_adjustment_workspace,
+                                   wavelength_and_pixel_adjustment_workspace=wavelength_and_pixel_adjustment_workspace)
+
             quadrant_scatter_reduced.update({quadrant: quadrant_scatter_data})
 
         # # ------------------------------------------------------------
@@ -368,7 +372,7 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         wavelength_adjustment = adjustment_alg.getProperty("OutputWorkspaceWavelengthAdjustment").value
         pixel_adjustment = adjustment_alg.getProperty("OutputWorkspacePixelAdjustment").value
         wavelength_and_pixel_adjustment = adjustment_alg.getProperty(
-                                           "OutputWorkspaceWavelengthAndPixelAdjustment").value
+            "OutputWorkspaceWavelengthAndPixelAdjustment").value
         return wavelength_adjustment, pixel_adjustment, wavelength_and_pixel_adjustment
 
     def _convert_to_histogram(self, workspace):
@@ -385,7 +389,7 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
 
         return workspace
 
-    def _convert_to_q(self, state_serialized, workspace, wavelength_adjustment_workspace, pixel_adjustment_workspace,
+    def _convert_to_q(self, state, workspace, wavelength_adjustment_workspace, pixel_adjustment_workspace,
                       wavelength_and_pixel_adjustment_workspace):
         """
         A conversion to momentum transfer is performed in this step.
@@ -399,24 +403,17 @@ class SANSBeamCentreFinderCore(DataProcessorAlgorithm):
         @param wavelength_and_pixel_adjustment_workspace: the wavelength and pixel adjustment workspace.
         @return: a reduced workspace
         """
-        convert_name = "SANSConvertToQ"
-        convert_options = {"InputWorkspace": workspace,
-                           "OutputWorkspace": EMPTY_NAME,
-                           "SANSState": state_serialized,
-                           "OutputParts": True}
-        if wavelength_adjustment_workspace:
-            convert_options.update({"InputWorkspaceWavelengthAdjustment": wavelength_adjustment_workspace})
-        if pixel_adjustment_workspace:
-            convert_options.update({"InputWorkspacePixelAdjustment": pixel_adjustment_workspace})
-        if wavelength_and_pixel_adjustment_workspace:
-            convert_options.update({"InputWorkspaceWavelengthAndPixelAdjustment":
-                                    wavelength_and_pixel_adjustment_workspace})
-        convert_alg = create_child_algorithm(self, convert_name, **convert_options)
-        convert_alg.execute()
-        data_workspace = convert_alg.getProperty("OutputWorkspace").value
-        sum_of_counts = convert_alg.getProperty("SumOfCounts").value
-        sum_of_norms = convert_alg.getProperty("SumOfNormFactors").value
-        return data_workspace, sum_of_counts, sum_of_norms
+        alg = ConvertToQ(state_convert_to_q=state.convert_to_q,
+                         wavelength_adj_workspace=wavelength_adjustment_workspace,
+                         pixel_adj_workspace=pixel_adjustment_workspace,
+                         wavelength_and_pixel_adj_workspace=wavelength_and_pixel_adjustment_workspace)
+        output_dict = alg.convert_workspace(workspace=workspace, output_summed_parts=True)
+
+        output_workspace = output_dict["output"]
+        sum_of_counts = output_dict["counts_summed"]
+        sum_of_norms = output_dict["norm_summed"]
+
+        return output_workspace, sum_of_counts, sum_of_norms
 
     def validateInputs(self):
         errors = dict()
