+++ Framework/PythonInterface/test/python/plugins/algorithms/WorkflowAlgorithms/sans/CMakeLists.txt
@@ -2,7 +2,6 @@
 
 set(TEST_PY_FILES
     SANSCalculateTransmissionTest.py
-    SANSConvertToQTest.py
     SANSConvertToWavelengthTest.py
     SANSCreateAdjustmentWorkspacesTest.py
     SANSCreateWavelengthAndPixelAdjustmentTest.py
